+++ exchange/exchangeChecker.js
