+++ exchange/exchangeErrors.js
