+++ exchange/portfolioManager.js
