+++ exchange/wrappers/bitstamp.js
