+++ exchange/wrappers/kraken-markets.json
