+++ exchange/wrappers/kraken.js
