+++ exchange/wrappers/mtgox.js
